+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 ## [9.11.0]
 ### Changed
 - Update SSDK 20.1.0 and BBOM 2026.03 LTS, boat-maven-plugin unified 0.17.66
+## [9.12.0]
+### Added
+- Implement Portfolio Risk assessments data ingestion;
+### Changed
+- use latest investment service api 1.4.1;
+- update investment data integraton service
 
 ## [9.10.0]
 ### Changed